# api.yaml
openapi: 3.0.0
info:
  title: Placement Portal API
  description: API specification for the Institute Placement Portal built using Flask.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local development server
paths:
...
  /admin/approve-company/{company_id}:
    post:
      summary: Approve company registration
      parameters:
        - name: company_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company approved
  /admin/approve-drive/{drive_id}:
    post:
      summary: Approve placement drive
      parameters:
        - name: drive_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Placement drive approved
...